BoringSSL tests: load ECH config
This implements a few of the options for running the BoringSSL tests for ECH. A few
checks are still disabled since I'm not sure how to get the required information
from the current API. Also the DTLS-ECH tests remain disabled until we get DTLS1.3.

I noticed two things that are harder than I feel they should be with the current API.
Retrieving the status requires assigning and freeing two variables, it would be a lot
more convenient if there's a method to get just the status and if the current one
would accept NULL to skip either of the outputs.

The same applies for setting the config. I've written quite some logic to get the
base64 config input into PEM-format in a BIO. In particular since I've got a single
config and need to transform it into a list.

Running the client tests shows quite some errors related to old versions in the
supported versions of the inner CH as well as TLS1.2 only extensions in the inner CH.

Added C++ header guards to <openssl/ech.h> as well.

[skip ci]