BoringSSL tests: implement related new options
Implement a number of new options related to options that already exist.

Repair session behaviour by duplicating the session stored in the TestState
in the NewSessionCallback. SSL_CTX_free will set the resume flag to 0 on its
sessions before freeing them. As a result, the stored session can't be used
for resumption unless it's duplicated.